refactor: rename package to android_file_picker while keeping folder structure
