+++ .github/dependabot.yml
@@ -30,7 +30,7 @@ updates:
           - "*"
 
   - package-ecosystem: "pub"
-    directory: "/packages/android_file_picker"
+    directory: "/packages/file_picker_android"
     schedule:
       interval: "weekly"
     groups:
@@ -85,7 +85,7 @@ updates:
 
   # Gradle dependencies for Android native plugin
   - package-ecosystem: "gradle"
-    directory: "/packages/android_file_picker/android"
+    directory: "/packages/file_picker_android/android"
     schedule:
       interval: "weekly"
     groups:
