+++ example/lib/src/file_picker_demo.dart
@@ -2,7 +2,7 @@
 
 import 'package:file/local.dart';
 import 'package:file_picker/file_picker.dart';
-import 'package:android_file_picker/android_file_picker.dart';
+import 'package:android_file_picker/file_picker_android.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
