+++ packages/file_picker_android/test/file_picker_android_test.dart
@@ -1,4 +1,4 @@
-import 'package:android_file_picker/android_file_picker.dart';
+import 'package:android_file_picker/file_picker_android.dart';
 import 'package:file_picker_platform_interface/file_picker_platform_interface.dart';
 import 'package:flutter_test/flutter_test.dart';
